pyproject.toml
[project]
name = "research-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    # Runtime config (used by agent/core_agent/config.py)
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.1",
]

[dependency-groups]
ai-agent = [
    "google-cloud-aiplatform[adk,agent-engines]>=1.138.0",
    "google-cloud-modelarmor>=0.4.0",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.1",
]
dev = [
    "google-cloud-modelarmor>=0.4.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.5",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.1",
    "click>=8.3.1",
    "dotenv>=0.9.9",
    "ipykernel>=7.2.0",
    "datetime>=6.0",
]
mcp_gcs = [
    "fastapi>=0.109.0",
    "google-cloud-storage>=2.14.0",
    "mcp>=1.2.0",
    "pydantic>=2.5.3",
    "sse-starlette>=2.0.0",
    "uvicorn>=0.27.0",
]
mcp_bq = [
    "google-cloud-bigquery>=3.20.0",
    "mcp>=1.26.0",
    "pydantic>=2.5.3",
]
mcp_drive = [
    "google-api-python-client>=2.0.0",
    "google-auth>=2.0.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.0",
    "mcp>=1.26.0",
    "pydantic>=2.5.3",
    "pypdf>=4.0.0",
    "reportlab>=4.0.0",
    "uvicorn>=0.27.0",
]

[tool.pytest.ini_options]
pythonpath = [
  "."
]

